change: the edit-image.js file is divided into two modules
