+++ js/main.js
@@ -1,5 +1,6 @@
 import './modal.js';
 import './form-validate.js';
-import { updateScaleValue } from './edit-image.js';
+import { updateScaleValue } from './scale-image.js';
+import './filter-image.js';
 
 updateScaleValue();
